feat: introduce new api client based on netbox-community/go-netbox
The NetBox API client setup is moved to netbox/client to
keep the provider setup code clean. The `Config` struct
is passed to the provider functions containing both the
legacy- and new API client.
The version check in the provider setup is the only API
call move to the new client.
The only thing changed in the existing resources is
the type assertion to access the legacy client.